backend/stockproject/predictor/requirements.txt
# Core framework
Django>=4.2

# Machine Learning / Deep Learning
tensorflow>=2.15.0
scikit-learn>=1.3.0

# Data handling
numpy>=1.24.0
pandas>=2.1.0
pandas_ta>=0.3.14b0
yfinance>=0.2.30

# CORS support
django-cors-headers>=4.2.0

# REST Framework (if you're using APIs)
djangorestframework>=3.14.0

# HTTP requests (optional for APIs, scraping etc.)
requests>=2.31.0
